Fix "This Month" example group
The total_inventory and recent donations expectations seem to have been
incorrect since introduction in 4597e6b. They worked previously because
the original expectations looked for the expected values anywhere on the
page.

Note that this will still fail on dev boxes during January due to
https://github.com/rubyforgood/human-essentials/issues/2676
but works on CI due to 87039ac

That will be addressed separately.